otel-flask/requirements.txt
flask
redis
fakeredis
requests
opentelemetry-api==1.18.0
opentelemetry-distro==0.39b0
opentelemetry-exporter-otlp-proto-common==1.18.0
opentelemetry-exporter-otlp-proto-grpc==1.18.0
opentelemetry-exporter-otlp-proto-http==1.18.0
opentelemetry-instrumentation==0.39b0
opentelemetry-instrumentation-aws-lambda==0.39b0
opentelemetry-instrumentation-boto3sqs==0.39b0
opentelemetry-instrumentation-botocore==0.39b0
opentelemetry-instrumentation-celery==0.39b0
opentelemetry-instrumentation-dbapi==0.39b0
opentelemetry-instrumentation-flask==0.39b0
opentelemetry-instrumentation-grpc==0.39b0
opentelemetry-instrumentation-jinja2==0.39b0
opentelemetry-instrumentation-logging==0.39b0
opentelemetry-instrumentation-pymongo==0.39b0
opentelemetry-instrumentation-redis==0.39b0
opentelemetry-instrumentation-requests==0.39b0
opentelemetry-instrumentation-sqlalchemy==0.39b0
opentelemetry-instrumentation-sqlite3==0.39b0
opentelemetry-instrumentation-system-metrics==0.39b0
opentelemetry-instrumentation-tortoiseorm==0.39b0
opentelemetry-instrumentation-urllib==0.39b0
opentelemetry-instrumentation-urllib3==0.39b0
opentelemetry-instrumentation-wsgi==0.39b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
opentelemetry-util-http==0.39b0